Revert "[FIX] Redis Message DTO 추가  #91"

+++ src/main/java/com/example/weup/service/TodoService.java
@@ -12,7 +12,6 @@ import com.example.weup.repository.MemberRepository;
 import com.example.weup.repository.ProjectRepository;
 import com.example.weup.repository.TodoMemberRepository;
 import com.example.weup.repository.TodoRepository;
-import com.example.weup.validate.MemberValidator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -27,8 +26,8 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class TodoService {
 
+    private final MemberService memberService;
     private final S3Service s3Service;
-    private final MemberValidator memberValidator;
     private final ProjectRepository projectRepository;
     private final MemberRepository memberRepository;
     private final TodoRepository todoRepository;
@@ -37,7 +36,12 @@ public class TodoService {
     @Transactional
     public void createTodo(Long userId, CreateTodoRequestDTO createTodoRequestDTO) {
 
-        memberValidator.validateActiveMemberInProject(userId, createTodoRequestDTO.getProjectId());
+        Member requestMember = memberRepository.findByUser_UserIdAndProject_ProjectId(userId, createTodoRequestDTO.getProjectId())
+                .orElseThrow(() -> new GeneralException(ErrorInfo.NOT_IN_PROJECT));
+
+        if (memberService.isDeletedMember(requestMember.getMemberId())) {
+            throw new GeneralException(ErrorInfo.DELETED_MEMBER);
+        }
 
         Project project = projectRepository.findById(createTodoRequestDTO.getProjectId())
                 .orElseThrow(() -> new GeneralException(ErrorInfo.PROJECT_NOT_FOUND));
@@ -51,7 +55,12 @@ public class TodoService {
 
     public List<TodoListResponseDTO> getTodoList(Long userId, Long projectId) {
 
-        Member requestMember = memberValidator.validateActiveMemberInProject(userId, projectId);
+        Member requestMember = memberRepository.findByUser_UserIdAndProject_ProjectId(userId, projectId)
+                .orElseThrow(() -> new GeneralException(ErrorInfo.NOT_IN_PROJECT));
+
+        if (memberService.isDeletedMember(requestMember.getMemberId())) {
+            throw new GeneralException(ErrorInfo.DELETED_MEMBER);
+        }
 
         List<Todo> todos = todoRepository.findByProject_ProjectId(projectId);
 
@@ -101,9 +110,21 @@ public class TodoService {
         Todo todo = todoRepository.findById(editTodoRequestDTO.getTodoId())
                 .orElseThrow(() -> new GeneralException(ErrorInfo.TODO_NOT_FOUND));
 
-        memberValidator.validateActiveMemberInProject(userId, todo.getProject().getProjectId());
+        Member requestMember = memberRepository.findByUser_UserIdAndProject_ProjectId(userId, todo.getProject().getProjectId())
+                .orElseThrow(() -> new GeneralException(ErrorInfo.NOT_IN_PROJECT));
 
-        todo.edit(editTodoRequestDTO.getTodoName(), editTodoRequestDTO.getStartDate(), editTodoRequestDTO.getEndDate());
+        if (memberService.isDeletedMember(requestMember.getMemberId())) {
+            throw new GeneralException(ErrorInfo.DELETED_MEMBER);
+        }
+
+        if (editTodoRequestDTO.getTodoName() != null) {
+            todo.setTodoName(editTodoRequestDTO.getTodoName());
+        }
+
+        if (editTodoRequestDTO.getStartDate() != null && editTodoRequestDTO.getEndDate() != null) {
+            todo.setStartDate(editTodoRequestDTO.getStartDate());
+            todo.setEndDate(editTodoRequestDTO.getEndDate());
+        }
 
         todoRepository.save(todo);
 
@@ -134,9 +155,14 @@ public class TodoService {
         Todo todo = todoRepository.findById(editTodoStatusRequestDTO.getTodoId())
                 .orElseThrow(() -> new GeneralException(ErrorInfo.TODO_NOT_FOUND));
 
-        memberValidator.validateActiveMemberInProject(userId, todo.getProject().getProjectId());
+        Member requestMember = memberRepository.findByUser_UserIdAndProject_ProjectId(userId, todo.getProject().getProjectId())
+                .orElseThrow(() -> new GeneralException(ErrorInfo.NOT_IN_PROJECT));
 
-        todo.changeStatus(editTodoStatusRequestDTO.getStatus());
+        if (memberService.isDeletedMember(requestMember.getMemberId())) {
+            throw new GeneralException(ErrorInfo.DELETED_MEMBER);
+        }
+
+        todo.setTodoStatus(editTodoStatusRequestDTO.getStatus());
         todoRepository.save(todo);
     }
 
@@ -146,7 +172,12 @@ public class TodoService {
         Todo todo = todoRepository.findById(todoId)
                 .orElseThrow(() -> new GeneralException(ErrorInfo.TODO_NOT_FOUND));
 
-        memberValidator.validateActiveMemberInProject(userId, todo.getProject().getProjectId());
+        Member requestMember = memberRepository.findByUser_UserIdAndProject_ProjectId(userId, todo.getProject().getProjectId())
+                .orElseThrow(() -> new GeneralException(ErrorInfo.NOT_IN_PROJECT));
+
+        if (memberService.isDeletedMember(requestMember.getMemberId())) {
+            throw new GeneralException(ErrorInfo.DELETED_MEMBER);
+        }
 
         todoMemberRepository.deleteAllByTodo_TodoId(todoId);
